api/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
psycopg[binary]
pydantic
python-dotenv
pandas
openpyxl
